+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 This file is for listing all the changes to this project
 
+## v1.1.0 Release | Nov 8, 2023
+### Adding
+- Progression Display `You can now pin any challenges you want on the screen.`
+- Adjustable Number of Challenges for each Tier Config `You can change the values on how many challenges should be given in each tier.`
+  *But remember that you must have the exact number of challenges in the Datapack. Otherwise, it will make the game crash!*
+### Fixed
+- [Forge] Saving player ability after respawn
+
+## v0.1.1a Beta | Oct 13, 2023
+### Fixed
+- [Forge] issue within OnDatapackSyncEvent method
+
 ## v0.1.1 Beta | Oct 11, 2023
 - Fixing issue related to packet